Move `find_all_ret_expressions` into `utils`
